+++ resources/views/posts/show.blade.php
@@ -7,8 +7,11 @@
     <hr>
     <div>{{ $post->body }}</div>
     <hr>
-    <a href="/posts/{{ $post->id }}/edit" class="btn btn-secondary">Edit</a>
-    {!! Form::open(['action' => ['PostsController@destroy', $post->id], 'method' => 'DELETE']) !!}
-        {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
-    {!! Form::close() !!}
+    @if (!Auth::guest() && (auth()->user()->id == $post->user_id ))
+        <a href="/posts/{{ $post->id }}/edit" class="btn btn-secondary">Edit</a>
+
+        {!! Form::open(['action' => ['PostsController@destroy', $post->id], 'method' => 'DELETE']) !!}
+            {{ Form::submit('Delete', ['class' => 'btn btn-danger']) }}
+        {!! Form::close() !!}
+    @endif
 @endsection
